Rename seqWindow into posBuffer
The term seqWindow is not correct because it manages more than the
window. We are using now the term pos-buffer to highlight the fact that
it also managed the position of the buffer in the byte stream.

We use s as receiver name for posBuffer because the buffer will be
embedded into sequencer types and using s as the name of the receiver
ensures that methods look not out of place in the Go documentation by
using another receiver name.