Rename HashSequencerConfig into HSConfig
With expected sequencer names like OptimalSuffixArraySequencer the names
for the configuration parameters would get too long. Therefore we will
abbreviate the name of the configuration types. For the HashSequencer it
is HSConfig. The name is unique and has a clear relationship with
HashSequencer. It is also short, which makes it easy to use.