Add data argument to Reset method
For the use of parallel Goroutines we want to be able to update the
window of the sequencers with the new data. For this we have to rewrite
the Reset method of the sequencer. Since the new data slice might be
larger than the window size we need to be able to return errors from the
sequencer.