lz: cleanup
The current implementation had issues. The window contained more than
the actual window. I renamed it SeqBuffer. The Buffer structure could
only be used for decoding. It is renamed to DecBuffer. By letting the
configuration types to embed SBConfig the configuration of the sequence
buffer, we could simplify the boiler plate for the sequencer
implementations.

We removed the ring buffer because it is slower than the DecBuffer.