Improve SeqBuffer configuration
The LZMA compression requires the full window to be present for
compression because the literal encoder may access the complete window.
So we need set ShrinkSize equal to WindowSize, which didn't work with
the current SBConfig verifier. This lead to a complete a review of the
relationship between ShrinkSize, BufferSize and WindowSize. ShrinkSize
and WindowSize are independent, but BufferSize must be larger than the
ShrinkSize and larger or equal to the WindowSize.