btree: create adapt function
The adapt function moves the pointers stored in the B-Tree using a
non-negative delta value. It is assumed that the content of the slice p
has been moved before calling adapt.

The commit renames the shift method to adapt and removed the copy call,
which shifted data inside the p slice.