hs.go: fix a bug found by fuzzing
FuzzHashSequencer found an issue. The HashSequencer didn't find a single
sequence during the run and didn't add literals because of the
flags the wrapper provides, which prevented wrapper from making progress
and ended in a loop.

I have decided now that trailing literals must actually trail a sequence
to be removed from the block. The other fuzz cases are simply testing
the verifier and are not so critical.