Use more sensible names.
We adapt our terminology to that of the literature. The process to
generate the LZ77 sequences is parsing. So the former sequencers have to
be named parsers. The SeqBuffer becomes new a ParserBuffer and the
DecBuffer a DecoderBuffer. Finders become dictionaries. The code is now
better to read because it uses familiar terminology.