Simplified API
I solved the WindowSize issue by introducing a generic function that
uses reflection to retrieve and set the WindowSize using recursive
lookup for WindowSize field of Configurator.

Using this approach the Configurator interfaces know only the New
method and stay simple. It also removes all the functions for member
access for the actual option structures.